+++ examples/middleware/http/gin/main.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 
 	"github.com/aserto-dev/go-aserto"
-	"github.com/aserto-dev/go-aserto/authorizer"
+	"github.com/aserto-dev/go-aserto/az"
 	"github.com/aserto-dev/go-aserto/middleware"
 	"github.com/aserto-dev/go-aserto/middleware/http/ginz"
 	"github.com/gin-gonic/gin"
@@ -15,7 +15,7 @@ import (
 const port = 8080
 
 func main() {
-	azClient, err := authorizer.New(
+	azClient, err := az.New(
 		aserto.WithAddr("localhost:8282"),
 	)
 	if err != nil {
