+++ examples/middleware/http/gorilla/main.go
@@ -8,7 +8,7 @@ import (
 	"github.com/gorilla/mux"
 
 	"github.com/aserto-dev/go-aserto"
-	"github.com/aserto-dev/go-aserto/authorizer"
+	"github.com/aserto-dev/go-aserto/az"
 	"github.com/aserto-dev/go-aserto/middleware"
 	"github.com/aserto-dev/go-aserto/middleware/http/std"
 )
@@ -16,7 +16,7 @@ import (
 const port = 8080
 
 func main() {
-	azClient, err := authorizer.New(
+	azClient, err := az.New(
 		aserto.WithAddr("localhost:8282"),
 	)
 	if err != nil {
